Create Drunken_Sailor_Survival_Model
A jetty that is 10 meters wide and 50 meters long. A Sailor staggers from the middle of one end of the jetty towards her ship moored at the other end. She always faces her ship, and at each step she has a probability p of taking a step forward, and a probability of q of lurching left or right, where q = (1 / p) / 2. Each step is 0.25 meters long (shes that a very short step...recall this is a very drunk sailor). If she falls off the jetty she drowns, but if she reaches the end her shipmates haul her on board. If she does not reach the ship after 2000 steps she dies of alcohol poisoning. 

I used a random number generator to simulate the sailor's progress, given probability p of moving forwards, and hence estimated the probability s that she reaches her ship safely. I provided the plot of the probability s of reaching the ship as a function of probability p of moving forwards.